// frontend/src/common/models/Crash.ts
export interface Crash {
    crasH_DATE: string;
  crasH_DAY_OF_WEEK: string;
  crasH_HOUR: string;
  crasH_MONTH: string;
  crasH_RECORD_ID: string;
  crasH_TYPE: string;
  firsT_CRASH_TYPE: string;
  hiT_AND_RUN_I: string;
  injurieS_TOTAL: string;
  lanE_CNT: string;
  latitude: string;
  longitude: string;
  posteD_SPEED_LIMIT: string;
  reporT_TYPE: string;
  streeT_DIRECTION: string;
  streeT_NAME: string;
  traffiC_CONTROL_DEVICE: string;
  weatheR_CONDITION: string;
}